updated test utils and entities
